Rename module from vo to vo_service_discovery
