+++ astroquery/vo_service_discovery/tests/data/registry_query_basic_result.ecsv
