+++ astroquery/vo_service_discovery/tests/setup_package.py
@@ -4,4 +4,4 @@ import os
 def get_package_data():
     paths_test = [os.path.join('data', '*.xml')] + [os.path.join('data', '*.ecsv')]
 
-    return {'astroquery.vo.tests': paths_test}
+    return {'astroquery.vo_service_discovery.tests': paths_test}
