+++ astroquery/vo_service_discovery/tests/test_registry_remote.py
